This is the cocaine of catergorical update checking.
It will show you what anitya thinks is the latest version for everything
in dev-perl/* , paginated, and/or tell you what's missing.

This is somehow faster than doing the same thing with any other API,
and doesn't rely on any weird ebuild-parsing magic trying to determine
the upstream tracking identifier, because that data is outsourced
to anitya.

Also means it *actually works* when upstream *ISNT* on CPAN!